In no_std context, use no_std num-traits.
This crate was not actually no_std compatible because it (1) used
num-traits with the `std` feature, and (2) relied on the `Float` trait,
which is unavailable in no_std.

I have altered it to rely on `FloatCore` instead and passed the feature
flag through to the dependency.

Had to bump the version of num-traits required, because FloatCore only
appeared in 0.2.1.