Document that `NotNan` is `repr(transparent)`.
Currently, rustdoc only displays `repr(transparent)` if the type’s
fields are public, so there is no indication in the documentation that
`NotNan` has `repr(transparent)`. Even if it was visible, there is not
currently a strong consensus that such a repr may be assumed to be a
stable part of the public API.

Therefore, add paragraphs to `NotNan` and `OrderedFloat`’s documentation
guaranteeing the repr. I also mentioned `bytemuck` so as to not be
encouraging unsafe code, though the phrasing isn’t great.